More code tidying
* Remove unnecessary escapes in regexes
* Remove unnecessary type casting
* Remove unnecessary StringBuilding
* General semi-automated code deduplication and simplification

git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20599 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/maximizer/Maximizer.java
@@ -1891,7 +1891,7 @@ public class Maximizer
 							text = "make & " + text;
 							cmd = "make \u00B6" + itemId + ";" + cmd;
 							price = ConcoctionPool.get( item ).price;
-							advCost = (int)ConcoctionPool.get( item ).getAdventuresNeeded( 1 );
+							advCost = ConcoctionPool.get( item ).getAdventuresNeeded( 1 );
 						}
 						else if ( checkedItem.npcBuyable > 0 )
 						{
